MDL-83243 core: User must leave page after two failed keepalive attempts
On pages using the session manager's keepalive logic, an unstable network
connection should cause an alert modal to appear warning the user that
their learning progress may not be saved. This doesn't always happen
because the modal contents must be fetched with a network request. This
commit pre-fetches the alert modal on page load to ensure that it is
available if network connection is lost, and if two consecutive requests
to touch the session fail, the modal will force the user to leave the
page, either by reloading or redirection.